demos moved to separate folder
