+++ src/app/demos/custom-template/custom-template.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { AppService } from '../app.service';
+import { AppService } from '../../app.service';
 import { FormControl } from '@angular/forms';
 import { Subscription } from 'rxjs';
