+++ src/app/demos/multiple-select-with-observable/multiple-select-with-observable.component.spec.ts
