+++ src/app/demos/multiple-select-with-observable/multiple-select-with-observable.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { AppService } from '../app.service';
+import { AppService } from '../../app.service';
 import { FormControl } from '@angular/forms';
 
 @Component({
