+++ src/app/demos/single-select-template-driven/single-select-template-driven.component.html
