+++ src/app/demos/single-select-with-property-map/single-select-with-property-map.component.html
