+++ src/app/title/title.component.html
@@ -1,5 +1,5 @@
 <mat-toolbar color="primary">
     <mat-toolbar-row>
-        <div>{{title}}</div>
+        <div>{{title}} this is title</div>
     </mat-toolbar-row>
 </mat-toolbar>
